+++ packages/runtime-client/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @sealant/runtime-client
 
+## 0.2.0
+
+### Minor Changes
+
+- 2861aaf: add SealantClient.fromStream for non-socket transports
+
+### Patch Changes
+
+- @sealant/runtime-protocol@0.2.0
+
 ## 0.1.3
 
 ### Patch Changes