+++ packages/runtime-protocol/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @sealant/runtime-protocol
 
+## 0.2.0
+
 ## 0.1.3
 
 ### Patch Changes